Split `report` and `benchmark` in two separate files
